Refactor to use protocol
