+++ biscoitometro/biscoitometro/Entities/Artist/ArtistDecodable.swift
@@ -7,3 +7,16 @@
 //
 
 import Foundation
+
+class ArtistDecodable: ArtistProtocol, Decodable {
+    let name: String
+    let id: Int
+    let profilePath: String?
+    
+    private enum CodingKeys: String, CodingKey {
+        case name = "name"
+        case id = "id"
+        case profilePath = "profile_path"
+    }
+}
+
