+++ biscoitometro/biscoitometro/Network Responses/SearchMovieResult.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct SearchMovieResult: Decodable, Paginable {
-    let results: [Movie]
+    let results: [MovieDecodable]
     var currentPage: Int
     var totalPages: Int
