+++ biscoitometro/biscoitometro/biscoitometro.xcdatamodeld/biscoitometro.xcdatamodel/contents
@@ -1,21 +1,21 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="14903" systemVersion="18G95" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="Artist" representedClassName=".Artist" syncable="YES">
+    <entity name="ArtistStorable" representedClassName=".ArtistStorable" syncable="YES">
         <attribute name="id" optional="YES" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
         <attribute name="name" optional="YES" attributeType="String"/>
         <attribute name="profilePath" optional="YES" attributeType="String"/>
-        <relationship name="castedIn" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="Movie" inverseName="casts" inverseEntity="Movie"/>
+        <relationship name="castedIn" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="MovieStorable" inverseName="casts" inverseEntity="MovieStorable"/>
     </entity>
-    <entity name="Movie" representedClassName=".Movie" syncable="YES">
+    <entity name="MovieStorable" representedClassName=".MovieStorable" syncable="YES">
         <attribute name="backdropPath" optional="YES" attributeType="String"/>
         <attribute name="id" optional="YES" attributeType="Integer 64" defaultValueString="0" usesScalarValueType="YES"/>
         <attribute name="overview" optional="YES" attributeType="String"/>
         <attribute name="posterPath" optional="YES" attributeType="String"/>
         <attribute name="title" optional="YES" attributeType="String"/>
-        <relationship name="casts" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="Artist" inverseName="castedIn" inverseEntity="Artist"/>
+        <relationship name="casts" optional="YES" toMany="YES" deletionRule="Nullify" destinationEntity="ArtistStorable" inverseName="castedIn" inverseEntity="ArtistStorable"/>
     </entity>
     <elements>
-        <element name="Artist" positionX="-63" positionY="-18" width="128" height="103"/>
-        <element name="Movie" positionX="-54" positionY="-9" width="128" height="133"/>
+        <element name="ArtistStorable" positionX="-63" positionY="-18" width="128" height="103"/>
+        <element name="MovieStorable" positionX="-54" positionY="-9" width="128" height="133"/>
     </elements>
 </model>
